pyproject.toml
[project]
name = "new"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "apify-client>=1.12.0",
    "google-generativeai>=0.8.5",
    "mcp[cli]>=1.11.0",
    "pymupdf>=1.26.3",
    "python-dotenv>=1.1.1",
    "streamlit>=1.47.0",
]
